refactor tests to match deep_visit changes
